requirements.txt
groq
fastapi
uvicorn[standard]
pandas
openpyxl
pdfplumber
tavily-python
python-dotenv
aiofiles
python-multipart
scikit-learn
numpy
voyageai
mcp
langsmith
psycopg2-binary
sqlalchemy
